+++ markwon-core/src/main/java/io/noties/markwon/LinkResolverDef.java
@@ -5,15 +5,20 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.provider.Browser;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 
 import androidx.annotation.NonNull;
 
 public class LinkResolverDef implements LinkResolver {
+
+    // @since 4.3.0-SNAPSHOT
+    private static final String DEFAULT_SCHEME = "https";
+
     @Override
     public void resolve(@NonNull View view, @NonNull String link) {
-        final Uri uri = Uri.parse(link);
+        final Uri uri = parseLink(link);
         final Context context = view.getContext();
         final Intent intent = new Intent(Intent.ACTION_VIEW, uri);
         intent.putExtra(Browser.EXTRA_APPLICATION_ID, context.getPackageName());
@@ -22,5 +27,19 @@ public class LinkResolverDef implements LinkResolver {
         } catch (ActivityNotFoundException e) {
             Log.w("LinkResolverDef", "Actvity was not found for the link: '" + link + "'");
         }
+    }
+
+    /**
+     * @since 4.3.0-SNAPSHOT
+     */
+    @NonNull
+    private static Uri parseLink(@NonNull String link) {
+        final Uri uri = Uri.parse(link);
+        if (TextUtils.isEmpty(uri.getScheme())) {
+            return uri.buildUpon()
+                    .scheme(DEFAULT_SCHEME)
+                    .build();
+        }
+        return uri;
     }
 }